+++ src/Savegame.h
@@ -26,7 +26,7 @@ struct SaveGame {
     size_t dungeonWidth = 0;
     size_t dungeonHeight = 0;
     size_t startingRoomIndex = 0;
-    std::unordered_map<uint32_t, RoomData> DungeonRooms; // hash corresponds to room index
+    std::unordered_map<size_t, RoomData> DungeonRooms; // hash corresponds to room index
 };
 
 nlohmann::json writeDataToJSON(const SaveGame& saveGame);
